application/requirements.txt
autopep8==1.6.0
click==8.0.4
dnspython==2.2.1
email-validator==1.2.1
Flask==2.0.3
Flask-WTF==1.0.1
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
pycodestyle==2.8.0
python-dotenv==0.20.0
toml==0.10.2
Werkzeug==2.0.3
WTForms==3.0.1
